+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+
+[0.1.0] - add group of items
+
+[0.0.2] - add responsible for web / desktop
+
 [0.0.1] - First Release but with good things